+++ Libplanet/Net/Swarm.cs
@@ -49,7 +49,6 @@ namespace Libplanet.Net
         private readonly PrivateKey _privateKey;
         private readonly int _appProtocolVersion;
 
-        private readonly TimeSpan _dialTimeout;
         private readonly AsyncLock _runningMutex;
         private readonly AsyncLock _blockSyncMutex;
         private readonly string _host;
@@ -86,33 +85,6 @@ namespace Libplanet.Net
             BlockChain<T> blockChain,
             PrivateKey privateKey,
             int appProtocolVersion,
-            int millisecondsDialTimeout = 15000,
-            string host = null,
-            int? listenPort = null,
-            DateTimeOffset? createdAt = null,
-            IEnumerable<IceServer> iceServers = null,
-            EventHandler<DifferentProtocolVersionEventArgs>
-                differentVersionPeerEncountered = null)
-            : this(
-                  blockChain,
-                  privateKey,
-                  appProtocolVersion,
-                  TimeSpan.FromMilliseconds(millisecondsDialTimeout),
-                  null,
-                  null,
-                  host,
-                  listenPort,
-                  createdAt,
-                  iceServers,
-                  differentVersionPeerEncountered)
-        {
-        }
-
-        public Swarm(
-            BlockChain<T> blockChain,
-            PrivateKey privateKey,
-            int appProtocolVersion,
-            TimeSpan dialTimeout,
             string host = null,
             int? listenPort = null,
             DateTimeOffset? createdAt = null,
@@ -123,7 +95,6 @@ namespace Libplanet.Net
                 blockChain,
                 privateKey,
                 appProtocolVersion,
-                dialTimeout,
                 null,
                 null,
                 host,
@@ -138,7 +109,6 @@ namespace Libplanet.Net
             BlockChain<T> blockChain,
             PrivateKey privateKey,
             int appProtocolVersion,
-            TimeSpan dialTimeout,
             int? tableSize,
             int? bucketSize,
             string host = null,
@@ -152,7 +122,6 @@ namespace Libplanet.Net
 
             _blockChain = blockChain ?? throw new ArgumentNullException(nameof(blockChain));
             _privateKey = privateKey ?? throw new ArgumentNullException(nameof(privateKey));
-            _dialTimeout = dialTimeout;
             LastSeenTimestamps =
                 new ConcurrentDictionary<Peer, DateTimeOffset>();
 
@@ -355,18 +324,24 @@ namespace Libplanet.Net
         }
 
         public async Task StartAsync(
+            int millisecondsDialTimeout = 15000,
             int millisecondsBroadcastTxInterval = 5000,
             CancellationToken cancellationToken = default(CancellationToken))
         {
             await StartAsync(
+                TimeSpan.FromMilliseconds(millisecondsDialTimeout),
                 TimeSpan.FromMilliseconds(millisecondsBroadcastTxInterval),
                 cancellationToken
             );
         }
 
+#pragma warning disable MEN002 // Line is too long
         /// <summary>
         /// Starts to periodically synchronize the <see cref="BlockChain"/>.
         /// </summary>
+        /// <param name="dialTimeout">
+        /// A timeout value for dialing.
+        /// </param>
         /// <param name="broadcastTxInterval">The time period of exchange of staged transactions.
         /// </param>
         /// /// <param name="cancellationToken">
@@ -382,9 +357,11 @@ namespace Libplanet.Net
         /// so that there are a lot of calls to <see cref="IAction.Render"/> method in a short
         /// period of time.  This can lead a game startup slow.  If you want to omit rendering of
         /// these actions in the behind blocks use <see cref=
-        /// "PreloadAsync(IProgress{PreloadState}, IImmutableSet{Address}, CancellationToken)"
+        /// "PreloadAsync(TimeSpan?, IProgress{PreloadState}, IImmutableSet{Address}, CancellationToken)"
         /// /> method too.</remarks>
+#pragma warning restore MEN002 // Line is too long
         public async Task StartAsync(
+            TimeSpan dialTimeout,
             TimeSpan broadcastTxInterval,
             CancellationToken cancellationToken = default(CancellationToken))
         {
@@ -459,7 +436,11 @@ namespace Libplanet.Net
 
             using (await _runningMutex.LockAsync())
             {
-                await PreloadAsync(render: true, cancellationToken: _cancellationToken);
+                await PreloadAsync(
+                    dialTimeout: dialTimeout,
+                    render: true,
+                    cancellationToken: _cancellationToken
+                );
                 Running = true;
             }
 
@@ -570,6 +551,9 @@ namespace Libplanet.Net
         /// <summary>
         /// Preemptively downloads blocks from registered <see cref="Peer"/>s.
         /// </summary>
+        /// <param name="dialTimeout">
+        /// A timeout value for dialing.
+        /// </param>
         /// <param name="progress">
         /// An instance that receives progress updates for block downloads.
         /// </param>
@@ -592,9 +576,10 @@ namespace Libplanet.Net
         /// </returns>
         /// <remarks>This does not render downloaded <see cref="IAction"/>s, but fills states only.
         /// If you want to render all <see cref="IAction"/>s from the genesis block to the recent
-        /// blocks use <see cref="StartAsync(TimeSpan, CancellationToken)"/> method
+        /// blocks use <see cref="StartAsync(TimeSpan, TimeSpan, CancellationToken)"/> method
         /// instead.</remarks>
         public Task PreloadAsync(
+            TimeSpan? dialTimeout = null,
             IProgress<PreloadState> progress = null,
             IImmutableSet<Address> trustedStateValidators = null,
             CancellationToken cancellationToken = default(CancellationToken)
@@ -602,6 +587,7 @@ namespace Libplanet.Net
         {
             return PreloadAsync(
                 render: false,
+                dialTimeout: dialTimeout,
                 progress: progress,
                 trustedStateValidators: trustedStateValidators,
                 cancellationToken: cancellationToken
@@ -611,6 +597,7 @@ namespace Libplanet.Net
         // FIXME: It is not guaranteed that states will be reported in order. see issue #436, #430
         internal async Task PreloadAsync(
             bool render,
+            TimeSpan? dialTimeout = null,
             IProgress<PreloadState> progress = null,
             IImmutableSet<Address> trustedStateValidators = null,
             CancellationToken cancellationToken = default(CancellationToken)
@@ -622,7 +609,10 @@ namespace Libplanet.Net
             }
 
             IList<(BoundPeer, long?)> peersWithHeight =
-                await DialToExistingPeers(cancellationToken).Select(pp =>
+                await DialToExistingPeers(
+                    dialTimeout,
+                    cancellationToken
+                ).Select(pp =>
                     (pp.Item1, pp.Item2.TipIndex)
                 ).ToListAsync(cancellationToken);
 
@@ -1077,6 +1067,7 @@ namespace Libplanet.Net
         }
 
         private IAsyncEnumerable<(BoundPeer, Pong)> DialToExistingPeers(
+            TimeSpan? dialTimeout,
             CancellationToken cancellationToken)
         {
             return new AsyncEnumerable<(BoundPeer, Pong)>(async yield =>
@@ -1086,7 +1077,11 @@ namespace Libplanet.Net
                     try
                     {
                         Message reply = await SendMessageWithReplyAsync(
-                            peer, new Ping(), _dialTimeout, cancellationToken);
+                            peer,
+                            new Ping(),
+                            dialTimeout,
+                            cancellationToken
+                        );
                         if (reply is Pong pong)
                         {
                             await yield.ReturnAsync((peer, pong));