+++ CHANGES.md
@@ -28,6 +28,8 @@ To be released.
  -  Removed `BlockChain<T>.Transactions`.  [[#409], [#583]]
  -  Removed the `linger` parameter from the `Swarm<T>` constructor and added
     `waitFor` to `Swarm<T>.StopAsync()`.  [[#581]]
+ -  Removed the `dialTimeout` parameter from the `Swarm<T>` constructor and
+    added it to `Swarm<T>.PreloadAsync()` and `Swarm<T>.StartAsync()`.  [[#581]]
 
 ### Added interfaces